obd2_collector/requirements.txt
obd==0.7.2
paho-mqtt==1.6.1
pyserial==3.5 